+++ ui/lib/extensions/src/react/content-block/edit-content-block.tsx
